Update: Test suite for syslog input plugin receiving via UDP
